requirements.txt

fastapi
uvicorn[standard]
loguru
rich
python-dotenv
anthropic
tavily
requests
prometheus-client
pydantic
pyyaml
